changes: ssmsql changed to postgresql and docker modifications

+++ Api/Models/DTOs/NewStationDto.cs
@@ -7,8 +7,8 @@
         public string Name { get; set; }
         public string Osoite { get; set; }
         public string Adress { get; set; }
-        public float x { get; set; }
-        public float y { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
         public int Kapasiteet { get; set; }
     }
 }
